2020.03 Release Candidate
A branch 'release' will be created out of this commit.

The commits in the branch 'release' will be squashed out and
files having references to CUDA will be deleted from that
branch as Nvidia's CUDA 11 has been pushed further into
May 2020.

So note that the 'release' branch will not have the
commit/development history in master branch.